Fix: revert 잘못 눌러서 다시 커밋

+++ here-draw-ios/here-draw-ios/Extension/UIImage+Extension.swift
@@ -10,7 +10,7 @@ import UIKit
 extension UIImage {
     func scaledHeight(scaledWidth: CGFloat) -> CGFloat {
         let oldSize = self.size
-        let scaleFactor = scaledWidth / oldSize.width
-        return oldSize.height * scaleFactor
+        let scaleFactor = scaledWidth / CGFloat(oldSize.width)
+        return CGFloat(oldSize.height) * scaleFactor
     }
 }
